added overpower skill to solar ray bunch of other solar ray fixes bunch of spell cast fixes removed AbstractSpell#whenCasting method added logo

+++ src/main/java/com/ombremoon/spellbound/common/magic/api/AnimatedSpell.java
@@ -62,7 +62,7 @@ public abstract class AnimatedSpell extends AbstractSpell {
             return this;
         }
 
-        public Builder<T> baseDamage(int baseDamage) {
+        public Builder<T> baseDamage(float baseDamage) {
             this.baseDamage = baseDamage;
             return this;
         }
